+++ src/EMRTimesIterator.h
@@ -24,7 +24,7 @@ public:
 protected:
 	EMRTimeIntervals                 m_intervs;
     EMRTimeIntervals::const_iterator m_iinterv;
-    size_t                           m_id_idx{0};
+    uint64_t                           m_id_idx{0};
     uint64_t                         m_num_steps;
     vector<uint64_t>                 m_num_steps4id;
 };
@@ -57,7 +57,7 @@ inline bool EMRTimesIterator::begin()
 {
     m_iinterv = m_intervs.begin();
     if (m_iinterv < m_intervs.end()) {
-        size_t num_ids = g_db->num_ids();
+        uint64_t num_ids = g_db->num_ids();
         for (m_id_idx = 0; m_id_idx < num_ids; ++m_id_idx) {
             m_point.id = g_db->id(m_id_idx);
             if (g_db->is_in_subset(m_point.id)) {
